Move catalog config endpoint contributor to runtime service SPI
