+++ bom/build.gradle.kts
@@ -115,6 +115,7 @@ dependencies {
     api(project(":polaris-runtime-defaults"))
     api(project(":polaris-server"))
     api(project(":polaris-runtime-service"))
+    api(project(":polaris-runtime-service-spi"))
     api(project(":polaris-runtime-spark-tests"))
 
     api(project(":polaris-tests"))
