+++ gradle/projects.main.properties
@@ -26,6 +26,7 @@ polaris-api-management-service=api/management-service
 polaris-api-catalog-service=api/polaris-catalog-service
 polaris-runtime-defaults=runtime/defaults
 polaris-runtime-service=runtime/service
+polaris-runtime-service-spi=runtime/service-spi
 polaris-server=runtime/server
 polaris-distribution=runtime/distribution
 polaris-runtime-spark-tests=runtime/spark-tests
